denso_pfmea/backend/pyproject.toml
[project]
name = "denso-pfmea-backend"
version = "0.1.0"
description = "PFMEA AI Workflow Backend - LangGraph Server"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langgraph-sdk>=0.1.0",
    "langchain-core>=0.3.0",
    "google-genai>=1.44.0",
    "google-auth>=2.35.0",
    "vertexai>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "structlog>=24.0.0",
    "tenacity>=8.0.0",
    "PyYAML>=6.0.0",
    "jinja2>=3.0.0",
    "openpyxl>=3.1.5",
    "pandas>=2.0.0",
    "langchain-google-genai>=4.2.0",
    "PyJWT>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "pandas-stubs>=2.0.0",
    "types-PyYAML>=6.0.0",
    "types-openpyxl>=3.1.0",
]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["langgraph.templates.agent", "agent"]
[tool.setuptools.package-dir]
"langgraph.templates.agent" = "src/agent"
"agent" = "src/agent"

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.mypy]
strict = true
python_version = "3.11"

[[tool.mypy.overrides]]
module = [
    "src.common.bop.*",
    "src.services.pfmea_ai",
    "src.services.llm_batch_runner",
    "src.services.pfmea.mapping_jobs",
    "src.agent.auth",
]
ignore_errors = true

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
]
lint.ignore = [
    "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]

[dependency-groups]
dev = [
    "anyio>=4.7.0",
    "langgraph-cli[inmem]>=0.4.7",
    "mypy>=1.13.0",
    "pytest>=8.3.5",
    "ruff>=0.8.2",
]
